pyproject.toml
[project]
name = "glados-personal-assistant-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "atomic-agents>=2.7.1",
    "fastapi>=0.135.1",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "uvicorn>=0.41.0",
]
